update code quality
