+++ pom.xml
@@ -22,7 +22,7 @@
 				<configuration>
 					<archive>
 						<manifest>
-							<mainClass>qengine.program.teamEngine.Main</mainClass>
+							<mainClass>qengine.program.teamengine.Main</mainClass>
 						</manifest>
 					</archive>
 				</configuration>
