+++ src/main/java/qengine/program/teamengine/MainRDFHandler.java
@@ -1,14 +1,14 @@
-package qengine.program.teamEngine;
+package qengine.program.teamengine;
 
 import org.eclipse.rdf4j.model.Statement;
 import org.eclipse.rdf4j.rio.helpers.AbstractRDFHandler;
-import qengine.program.teamEngine.index.OPS.OPS;
-import qengine.program.teamEngine.index.OSP.OSP;
-import qengine.program.teamEngine.index.POS.POS;
-import qengine.program.teamEngine.index.PSO.PSO;
-import qengine.program.teamEngine.index.SOP.SOP;
-import qengine.program.teamEngine.index.SPO.SPO;
-import qengine.program.teamEngine.q1.Dictionary;
+import qengine.program.teamengine.index.OPS.OPS;
+import qengine.program.teamengine.index.OSP.OSP;
+import qengine.program.teamengine.index.POS.POS;
+import qengine.program.teamengine.index.PSO.PSO;
+import qengine.program.teamengine.index.SOP.SOP;
+import qengine.program.teamengine.index.SPO.SPO;
+import qengine.program.teamengine.q1.Dictionary;
 
 /**
  * Le RDFHandler intervient lors du parsing de données et permet d'appliquer un traitement pour chaque élément lu par le parseur.
@@ -42,7 +42,5 @@ public final class MainRDFHandler extends AbstractRDFHandler {
         SOP.getInstance().add(subject, predicate, object);
         SPO.getInstance().add(subject, predicate, object);
     }
-
-    ;
 
 }
