+++ src/main/java/qengine/program/teamengine/timers/Timer.java
@@ -1,4 +1,4 @@
-package qengine.program.teamEngine.timers;
+package qengine.program.teamengine.timers;
 
 public class Timer {
     private double time = 0;
