+++ src/main/java/qengine/program/teamengine/utils/Constants.java
@@ -1,4 +1,4 @@
-package qengine.program.teamEngine.utils;
+package qengine.program.teamengine.utils;
 
 public class Constants {
     public static final int CONVERT_NS_TO_SEC = 1000000000;
